Add [mybatis framework]
使用mybatis框架操作数据库，摒弃之前的SpringData Jpa

+++ community-management-system-be/src/main/java/com/example/csgs/bean/PageQuery.java
@@ -10,8 +10,8 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PageQuery<T> {
-    private int curPage;
-    private int totalPage;
-    private int totalSize;
+    private long curPage;
+    private long totalPage;
+    private long totalSize;
     private List<T> dataList;
 }
